+++ ya_httpx_client/service_base/__init__.py
@@ -1,0 +1,2 @@
+from .vpn import VPNServiceBase
+from .file_serialization import FileSerializationServiceBase
